requirements.txt
pytest==6.2.2
python-dotenv==0.19.2
stomp.py==7.0.0